refactor: 🔄 Footer: molecules => atoms
