+++ packages/blog2/src/components/atoms/Footer/index.tsx
